backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
httpx==0.27.0
python-socketio==5.11.2
redis==5.0.4
celery==5.4.0
Pillow==10.3.0
boto3==1.34.101
sendgrid==6.11.0
authlib==1.3.0
google-auth==2.29.0
stripe==9.5.0
pytest==8.2.0
pytest-asyncio==0.23.6
